打入日志点

+++ xiaoerke-order/xiaoerke-order-service/src/main/java/com/cxqm/xiaoerke/modules/order/service/impl/ConsultPhonePatientServiceImpl.java
@@ -30,6 +30,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.text.SimpleDateFormat;
 import java.util.*;
 
@@ -187,8 +188,8 @@ public class ConsultPhonePatientServiceImpl implements ConsultPhonePatientServic
             Map<String,Object> parameter = systemService.getWechatParameter();
             String token = (String)parameter.get("token");
             String url = ConstantUtil.TITAN_WEB_URL+"/titan/phoneConsult#/orderDetail"+(String) map.get("doctorId")+","+(Integer) map.get("orderId")+",phone";
-            PatientMsgTemplate.consultPhoneCancel2Wechat(doctorName,date, week, beginTime,(String) map.get("endTime"), (String) map.get("phone"),(String) map.get("orderNo"),(Float) map.get("price")+"",(String) map.get("openid"),token,url);
-            PatientMsgTemplate.consultPhoneRefund2Msg(doctorName, (Float) map.get("price")+"", (String) map.get("phone"),(String) map.get("date"), week, (String) map.get("beginTime"),(String) map.get("orderNo"));
+            PatientMsgTemplate.consultPhoneCancel2Wechat(doctorName,date, week, beginTime,(String) map.get("endTime"), (String) map.get("phone"),(String) map.get("orderNo"),(BigDecimal) map.get("price")+"",(String) map.get("openid"),token,url);
+            PatientMsgTemplate.consultPhoneRefund2Msg(doctorName, (BigDecimal) map.get("price")+"", (String) map.get("phone"),(String) map.get("date"), week, (String) map.get("beginTime"),(String) map.get("orderNo"));
 
             //医生-短信
             DoctorMsgTemplate.cancelDoctorPhoneConsult2Sms(doctorName, babyName, doctorPhone, date, week, beginTime);
